Strip trailing null bytes from hash
The buffer into which the hash is filled is the maximum length but
the actual hash may be shorter and the rest of the buffer will remain
filled with null bytes, and, possibly surprisingly, these aren't
removed by .decode.  This doesn't affect the verification against
the hash as back in C-land the first \0 terminates the string.

I found this while trying to save the hash in a database using Red,
where it confused the driver by ending the string at the first \0
before an end quote.